revert(exec): leave the crontab writers unbounded, and say why
b141016 bounded the two `crontab -` writers to match their readers. Three refuters
were then pointed at it, one per decision, each required to run a differential
control. Two of them landed, and the bound does not survive.

THE BOUND CAN DESTROY THE TABLE. WaitDelay bounds the DRAIN, and cutting the drain
calls closeDescriptors, which SIGPIPEs anything still holding the inherited pipes.
b141016 assumed such a descendant is always WEDGED. If it is still WORKING it is
killed mid-install. Measured on this frame, differential, unbounded arm correct in
both cases: with the bound the table was NEVER WRITTEN where the unbounded run
installed it in eight seconds, and with a payload past the 64 KiB pipe buffer the
bound truncated it to exactly 65536 bytes where the unbounded run delivered all
204843. A merely slow but finite drain was also reported as a failure over a table
that had been written correctly.

The worst case is the second writer. Its error is DISCARDED (migrateLegacyCronEntries
is void) and applyCronMode calls it to establish the cron fallback and then removes
the daemon unit unconditionally, so a bound killing the descendant there leaves the
host with no cron line AND no daemon, reporting nothing. A wedged descendant, which
is what the bound was for, at least hangs visibly.

TWO CLAIMS IN b141016 WERE WRONG. "The budget starts only once the child has EXITED,
so it cannot truncate the payload": Go documents the trigger as ctx-done OR Wait
observing the exit, whichever comes first, and more to the point the child's exit is
not the end of the operation. "The cost of propagating is one log line that nothing
reads": true at four of the five call frames, false at daemon_setup.go:142, where on
the CLI --daemon-setup path a false ErrWaitDelay flips the final footer from green to
yellow and prints a WARNINGS/ERRORS block over a run that fully succeeded.

WHAT DOES NOT CHANGE. The truncated-table shape is real and was independently
confirmed: a descendant holding the stdin read end leaves crontab installing a partial
table and exiting 0, indistinguishable at the call frame from a healthy run (same
error, same empty capture, same elapsed time). That is why bounding this frame safely
needs a read back through crontabReadLines rather than a timer, and it is recorded in
the comment. The reader/writer asymmetry CodeRabbit flagged is now a documented
decision instead of an omission: a reader loses nothing when its drain is cut.

Neither the stall nor the damage can arise on a stock host, where crontab is a single
setuid binary that forks nothing. Both need a crontab that leaves children behind.

THE TESTS ARE REPLACED, NOT ADAPTED. The third refuter ran seventeen mutations against
the four cases b141016 added and only one carried its weight. c3 could not fail at any
value of CommandWaitDelay, including 1ns, while its own comment claimed to guard that.
c2 asserted elapsed time alone and stayed green when the writer was gutted, when the
payload was emptied, when PATH was replaced so no descendant spawned, and when the call
under test was deleted outright. The four cases are gone.

The three that replace them assert on the DELIVERED TABLE, which is what those
mutations were escaping through. Re-adding the bound to either writer is caught by that
writer's own case, and all four escapes the refuter found (deleted call, PATH replace,
emptied stdin on each site) are now caught. The descendant's lifetime is derived from
the shipped CommandWaitDelay rather than hardcoded, so the cases keep meaning the same
thing if the constant moves.
